code/training/requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.2
pandas>=2.0
scikit-learn>=1.2
tqdm>=4.65
Pillow>=9.5
